Change Option fold/foldStrict to foldLazy/fold (and same for getOrElse)
- Closes #44
- We have decided that Option `fold` and `getOrElse` functions should take a strict
default value for the `None` case, and the `foldLazy`/`getOrElseLazy`
functions should be the lazy alternatives.  The rationale is that it's
more convenient to just construct and throw away the default value 99%
of the time.  The `Lazy` version of the function is available in cases
where the default value is expensive to construct.