Have `IO.catchError` handle more cases
`IO.catchError` only handled the `Throw` variant so if you had an error in an `Async` variant you wouldn't catch it. Now it handles all of the cases that it should.

The only question I have is whether or not `Map` and `FlatMap` should be piped through this as well. I tried it but the compiler yelled at me so I stopped.