Reorganization
* Move data/ modules up a level, and remove data/ folder
* Rename classes/ folder to extensions/
* Rename _Types.re files to _Instances.re because they mostly contain
typeclass instances, and some of the related functions
* Rename _Submodules.re files to _Extensions.re because we are kind of
settling in on the name `Extensions` for module functor type stuff.
* Rename some of the Foldable extensions from things like OfEq to
FoldableEqExtensions.  I was trying to go in the direction of the `Make`
name for these things, but it just felt really awkward.  Also, the
`OfEq` names might conflict if we were to `include` other similar
extension modules, so I settled on `FoldableEqExtensions` as a unique,
descriptive name.