Initial implementation of Map/Set
Fixes #70

This is an initial implementation of wrapping Map/Set to make it more
ergonomic for consumption from Relude and projects that use Relude.

This implementation uses a module functor that expects an `ORD` to be
passed which is then used to construct the Map/Set. It also rewraps
the `Belt.Map.*` and `Belt.Set.*` functions to be pipe last and to
normalize the naming to be consistent with Relude naming.