Fix incorrect type signature for `Map.groupBy`
Currently, the `Map.groupBy` wants a `Map.Comparable.t` instead of the `Map.key`. These types should be the same but that isn't explicit in the module definition and it's opaque to consumers so the type signature of this needs to switch to be explicitly the key of the map.