Add `Named` comparison operators for ORD typeclass
When using comparison operators for ORD typeclasses, it's somewhat
ambiguous which one is being compared to which.

For example if you want to see 2 is less than 3 then you have to know
that with using the `OrdExtensions` methods that 2 has to come before
3 in the arg order. Also, when partially applying or when using `|>`
things don't read very well (e.g. `3.0 |> Float.lessThan(2.0)`).

By using a named parameter, the thought is that this can help clear that
up by making it explicit what is being compared to (e.g.
`2.0 |> Float.Named.lessThan(~compareTo=3.0)`)

The new functions are in a submodule called `Named` which was the done
to keep backwards compatibility. I'm not 100% on the naming of things
but I wanted to open this and start a conversation about what would be a
way forward that everybody can agree on.