Add Set.update
When working with sets, sometimes you need to update a value in the set 
and not just add a value to it only when it doesn't exist previously.

This function will remove an entry from the set if it's there and will 
add the new version of it into the set afterwards.