Add type defs to internal Comparable module in Map/Set
When using maps and sets, the type definition for the internal comparable is obfuscated so it cannot be used in any other context (even if the types should line up). For example, if you try to use the key of a `String.Map` as a string you will get a type error.

Adding the type defs to the module definition is a simple fix that will allow this functionality.